Add files via upload
In this assignment, I extracted some essential economic indicators from provided data and I displayed these economic indicators in a Visual.

Gross domestic product (GDP) is a measure of the market value of all the final goods and services produced in a period. GDP is an indicator of how well the economy is doing. A drop in GDP indicates the economy is producing less; similarly an increase in GDP suggests the economy is performing better. In this lab, I examined how changes in GDP impact the unemployment rate in the US.